Reverting Services.Contracts: 3.2.0 Added event DataSyncEntityWasAssociated
