+++ src/Services.Contracts/Events/SynchronizedEntity/DataSyncEntityWasUpdated.cs
@@ -33,4 +33,30 @@ namespace Nexus.Link.Services.Contracts.Events.SynchronizedEntity
             FulcrumValidate.IsValidated(Key, $"{propertyPath}.{nameof(Key)}", nameof(Key), errorLocation);
         }
     }
+
+    public class SyncKey : IValidatable
+    {
+        /// <summary>
+        /// The name of the sync client
+        /// </summary>
+        public string ClientName { get; set; }
+
+        /// <summary>
+        /// The name of the entity.
+        /// </summary>
+        public string EntityName { get; set; }
+
+        /// <summary>
+        /// The id of the object that has been updated.
+        /// </summary>
+        public string Value { get; set; }
+
+        /// <inheritdoc />
+        public void Validate(string errorLocation, string propertyPath = "")
+        {
+            FulcrumValidate.IsNotNullOrWhiteSpace(ClientName, nameof(ClientName), errorLocation);
+            FulcrumValidate.IsNotNullOrWhiteSpace(EntityName, nameof(EntityName), errorLocation);
+            // Value can be null when creating objects
+        }
+    }
 }
