+++ src/Services.Contracts/Services.Contracts.csproj
@@ -8,7 +8,7 @@
 
   <PropertyGroup>
     <PackageId>Nexus.Link.Services.Contracts</PackageId>
-    <Version>3.2.0</Version>
+    <Version>3.1.1</Version>
     <PackageRequireLicenseAcceptance>false</PackageRequireLicenseAcceptance>
     <PackageTags>nexus;link</PackageTags>
     <Authors>XLENT Link</Authors>
@@ -18,7 +18,6 @@
     <Copyright>Copyright ©2019 Xlent Link AB</Copyright>
     <IncludeSymbols>true</IncludeSymbols>
     <PackageReleaseNotes>
-      3.2.0 Added event DataSyncEntityWasAssociated
       3.1.1 Fix: PackageProjectUrl for this nuget package
       3.1.0 DataSyncMemory now supports IReadChildrenWithPaging.
       3.0.0 IPublishable now inherits from IValidatable
